Added in all necessary changes to HLLocationManager for monitoring updates in the background. We are currently adding to an array that isn't persisted, so persistence will be needed at some point.
Added in a function for determining the house location once 100 CLLocation's have been added to the array. This number should be able to be modified.
Commented out previous implementation of only checking on certain dates, this can be done later in the filtering process if I build up my own object that includes the date and location data.