Just a small update...
Location manager now persists a dictionary, this dictionary uses the days for keys to store an array of x amount of locations that should be tracked under the criteria of time and day.
Removed a fair amount of commented code. Now have to work on the comparing of the location objects saved, only wanting to collect the ones that are within ten meters of each other.
Updated the storyboard etc to work on my iPad for testing purposes.