Changed to isNight check instead of isDay. We only want to save the user's location if it is night. Changed limit of location objects to 100.
When determining the house location, we get all the objects for each day and create a centre location.
Added in if the user opens the application when tracking is not happening, we should do a check for whether or not we have the right amount of days to determine the house location.